Enable compiler -Werror and -Wall on our webkit code for mac and linux.
This fixes some initialization order, removes unused varibles,
and fixes a couple signed/unsigned comparison errors.  Please
check the signed/unsigned cases.

We can't enable it for windows yet because MSVC complains
about headers pulled in from WebCore.  We'll have to wait
until the API is upstreamed.

BUG=21140

Review URL: http://codereview.chromium.org/214024


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@26728 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: be21555a228e74b1ad6f3314ae882cf21bc3c24e

+++ webkit/api/src/gtk/WebFontInfo.cpp
@@ -41,7 +41,7 @@ namespace WebKit {
 WebCString WebFontInfo::familyForChars(const WebUChar* characters, size_t numCharacters)
 {
     FcCharSet* cset = FcCharSetCreate();
-    for (int i = 0; i < numCharacters; ++i) {
+    for (size_t i = 0; i < numCharacters; ++i) {
         if (U16_IS_SURROGATE(characters[i])
          && U16_IS_SURROGATE_LEAD(characters[i])
          && i != numCharacters - 1
